requirements.txt
datasets>=2.14.0
pandas>=2.0.0
scikit-learn>=1.3.0
sklearn-crfsuite>=0.3.6
pyspellchecker>=0.7.3
emoji>=2.8.0
joblib>=1.3.0
